backend/requirements.txt
acres==0.5.0
aerich==0.9.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiomysql==0.2.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncclick==8.1.8
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.4
dictdiffer==0.9.0
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
fitz==0.0.1.dev2
frontend==0.0.3
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib-resources==6.5.2
iso8601==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
jiter==0.10.0
looseversion==1.3.0
lxml==6.0.0
mako==1.3.10
markupsafe==3.0.2
multidict==6.6.3
mysql-connector-python==9.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
openai==0.28.0
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pdfminer-six==20250506
propcache==0.3.2
prov==2.1.1
puremagic==1.29
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydot==4.0.1
pymupdf==1.26.1
pymysql==1.1.1
pyparsing==3.2.3
pypika-tortoise==0.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
requests==2.32.4
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
tortoise-orm==0.25.1
tqdm==4.67.1
traits==7.0.2
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
